Support `message_type` attribute for s3logging
It'd be really useful to be able to specify `message_type` attribute on
s3logging configuration.

Potential use case for this is eg. being able to set message_type to blank and
use format to write logs to s3 as a JSON payload (we can use format to build
JSON structured logs).

`message_type` support in the underlying go-fastly library was added earlier as
part of PR https://github.com/sethvargo/go-fastly/pull/44