Add support for Logentries logging
As per [1] and [2] Fastly supports Real-Time Log Streaming to
Logentries.  The `go-fastly` library has got support to configure
Logentries as a logging target.

This commit wires the `go-fastly` support to Terraform Fastly provider
so no we're able to configure Logentries as a target for Real-Time Log
Streaming.  Tests coverage and documentation included as well.

[1] https://docs.logentries.com/v1.0/docs/fastly/
[2] https://docs.fastly.com/guides/streaming-logs/log-streaming-logentries