Expand documentation of GCS secret key
Keys are particularly difficult to troubleshoot due to the sensitive nature of their values. Google Cloud Platform usually releases credentials in JSON strings whereas specifying them as an environment variable [requires unusual encodings](https://stackoverflow.com/a/36439943/91590). How to pass credentials in is out of scope for this documentation, but I'd like a more precise explanation of what is required.