Changing formatVersion in VCLLoggingAttributes to be a *uint
As some logging endpoints on go-fastly use uint and some *uint, and
VCLLoggingAttributes needs to cater for both.
The most generic and recommended way to express empty value is using nil
instead of zero.
For the Request objects that are still using uint, the provider
will set the value to 0 (zero) when the VCLLoggingAttributes.formatVersion
pointer is nil. This is compatible with the previous behaviour.

Request objects using *uint which were failing before (as zero was being
sent instead of nill) are now working again.