Changing ACL, Dictionary and Dynamic Snippets documentation example to use for_each attributes instead of for expressions.
In order to use versionless resources (i.e. ACL Entries, Dictionary Items, Dynamic Snippets Content), it's required to link the ID generated from the Fastly Service (fastly_service_v1) to the respective field in the versionless resource (i.e fastly_service_acl_entries, fastly_service_dictionary_items_v1, fastly_service_dynamic_snippet_content_v1).

The original documentation instructed to the use of for expressions to create this link. For example:

resource "fastly_service_v1" "myservice" {
  ...
  dictionary {
    name = "dict1"
  }

}
resource "fastly_service_dictionary_items_v1" "items" {
  service_id = fastly_service_v1.myservice.id
  dictionary_id = {for d in fastly_service_v1.myservice.dictionary : d.name => d.dictionary_id}["dict1"]
  ...
}

This seems to work for many scenarios, however, not for all.
It fails mainly when the resource fastly_service_v1 already exists and we add the acl block at the same time, we add fastly_service_acl_entries resources (The same occur for the other versionless resources).

The generated error is: Error: Invalid index

This seems to be due to the way Terraform evaluates Computed attributes inside Sets. When the resource doesn't exist, Terraform postpones the evaluation of the expression as the values are not yet known.
However, when the resource exists and the Set attribute is being modified, Terraform tries to evaluate the expression before the all values are known causing the error. This is likely to be an issue in Terraform.

When replacing the for expression with a for_each attribute, Terraform is able to resolve the values correctly. That is, Terraform detects that a value the computed id is still not available and postpones the expression evaluation.

Interestingly enough, if using the for expression with an if and lookup, the issue seems to also disappear. For example

lookup({for d in fastly_service_v1.myservice.dictionary : d.name => d.dictionary_id if d.name == "dict1"}, "dict1", "")

This approach however is not as clean as the for_each one.

For all those reasons, the documentation was updated to recommend the for_each approach.

A similar issue also happens to List types and current affects the WAF resource. (This issue was previously raised against Terraform - see hashicorp/terraform#23488)

For example:

resource "fastly_service_v1" "demo" {
  ...
  waf {
    response_object    = "WAF_Response"
  }

}

resource "fastly_service_waf_configuration" "waf" {
  waf_id = fastly_service_v1.demo.waf[0].waf_id
}

This will cause an error, when the waf block is added to an existing at the same at time that we add the fastly_service_waf_configuration

Unfortunately for case, there's currently no alternative approach to the index lookup. The for_each approach doesn't work correctly.

Therefore, WAF documentation was updated to include a warning about adding WAF to an existing service and a recommendation to do a two-phase approach in case this change is required.

Note:
For this PR, I've also tried to update the acceptance tests' HCL to reflect the for_each recommendation.
However, there’s apparently another issue in the Terraform Test framework that fails to destroy resources at the end of the test step if the HCL contains for_each. See related issue hashicorp/terraform-plugin-sdk#264.