Add force_destroy argument to allow opt-in to delete active subscription
Previously was always effectively "force destroying" by unconditionally
deleting all associated TLS activations. I have been introduced to the
"force" flag in the API which is a more elegant way to do the same
thing. Having added that to the go-fastly SDK, I am taking advantage of
that here, and also making it optional behaviour so that the user has to
choose to take the risk of disabling production traffic.