Reparallelise most splunk/syslog tests, having identified problem
Some of the tests (not all) set env variables to exercise the
schema.DefaultEnvFunc which means they can't be parallelised (global env
between tests). However, as long as the rest of the tests are run at a
different time, this isn't an issue. Therefore I have left only the
tests using env vars as serial and returned the others to using
ParallelTest.