Avoid d.Set panic by pruning superfluous map keys beforehand
A lot of the logging block tests for the compute service were panicking
with the new SDK due to d.Set errors. These errors were ignored in the
resource Read functions, which would previously not have caused an issue
but now result in panics. The error was caused by trying to set VCL
logging attributes on compute services when these fields aren't present
in the schema.

This commit avoids the error by pruning these fields on Compute services
prior to calling d.Set.