Test speedup (#371)
* Replace Test with ParallelTest

* Splunk and syslog back to serial tests due to TLS CA cert sharing

* Reparallelise most splunk/syslog tests, having identified problem

Some of the tests (not all) set env variables to exercise the
schema.DefaultEnvFunc which means they can't be parallelised (global env
between tests). However, as long as the rest of the tests are run at a
different time, this isn't an issue. Therefore I have left only the
tests using env vars as serial and returned the others to using
ParallelTest.

* Add comments explaining lack of parallel test

* Improve table-driven test comment explaining t.Parallel outside resource.Test

Co-authored-by: Mark McDonnell <Integralist@users.noreply.github.com>

* Configure test parallelism in Makefile

* Document test parallelism in README

* Document the fact that tests may fail due to unavailable features

Co-authored-by: Mark McDonnell <Integralist@users.noreply.github.com>