Add integration tests for ACL/Dict force_destroy behaviour
Had to use a resource.TestCheckFunc with a side effect to add an
entry/item as part of a TestStep. This leads to a six part test where we
test renaming an empty list, add an entry, try renaming and expect an
error, then use force_destroy and retry renaming, expecting it to
succeed.