SchemaValidateFunc -> SchemaValidateDiagFunc
Quite a bit of coercion to get everyone happy because the signature has changed quite a bit from:

  func(interface{}, string) ([]string, []error)

 to:

   func(interface{}, cty.Path) diag.Diagnostics

Which requires a diagsToWarnsAndErrs function and cty.GetAttrPath() in the unit tests for all of the validators.
The validators themselves use the validation.ToDiagFunc() function to do this conversion which is quite simple, and also required for all of the builtin validation helper methods which have weirdly not yet been updated to SchemaValidateDiagFuncs.