Terraform Plugin SDK Upgrade (#379)
* Random docs changes (maybe a tfplugindocs change?)

* Upgrade TF Plugin SDK to v2 (tests panic)

Initial upgrade to get everything compiling with the new version.
Doesn't tackle any deprecations and doesn't have a working test suite yet.

Changes are mostly a direct find/replace of the SDK imports, with some more significant changes:
- The UserAgent generation was changed to account for an SDK change (fastly/config.go)
- The hashcode package was removed from the SDK and it is needed in some of the data source code, so it has been copied back in locally (fastly/hashcode/hashcode.go)
- Removal of d.SetPartial and d.Partial as it's been deprecated and doesn't seem to do anything
- Addition of context to CustomizeDiffFunc as the signature has been updated in-place
- A couple of lint errors that came up while trying to commit

* Avoid d.Set panic by pruning superfluous map keys beforehand

A lot of the logging block tests for the compute service were panicking
with the new SDK due to d.Set errors. These errors were ignored in the
resource Read functions, which would previously not have caused an issue
but now result in panics. The error was caused by trying to set VCL
logging attributes on compute services when these fields aren't present
in the schema.

This commit avoids the error by pruning these fields on Compute services
prior to calling d.Set.

* Fix dynamicsnippet content test

Testing TypeSet elements has changed in acceptance tests. Looking up a
preknown hash in the set no longer works, but there are new
TestCheckTypeSetElemAttr and TestCheckTypeSetElemNestedAttrs functions
for this instead which use "set.*" as the syntax.

Also the test was working incorrectly with a "count" attribute. I'm
guessing this worked before because a fake terraform implementation was
being used, but I couldn't get this working now. I tried ....content[0]
to try indexing the first instance of the "count"ed resource but it
still didn't work. In the end, I noticed that it was a bit redundant
anyway because there was always one thing being created, and the rest of
the fields would all have been duplicated if there were multiple.

* Add note about dropped TF 0.11.0 support

* Change CRUD functions to new context-aware type and use diag.Diagnostics

Subscription Validation resource not compiling as I need to change the retry mechanism

* Add diagToErr function to convert diag.Diagnostics to a Go error type

Used twice, once in the Service State Importer function, and once in the Subscription Validation Retry function. These are both context-aware but don't use diag.Diagnostics so need some conversion in order to work with the context-aware resource.*Read functions.

* Remove redundant HTTPS URL validator

* SchemaValidateFunc -> SchemaValidateDiagFunc

Quite a bit of coercion to get everyone happy because the signature has changed quite a bit from:

  func(interface{}, string) ([]string, []error)

 to:

   func(interface{}, cty.Path) diag.Diagnostics

Which requires a diagsToWarnsAndErrs function and cty.GetAttrPath() in the unit tests for all of the validators.
The validators themselves use the validation.ToDiagFunc() function to do this conversion which is quite simple, and also required for all of the builtin validation helper methods which have weirdly not yet been updated to SchemaValidateDiagFuncs.

* State -> StateContext

* Find potentially uncaught error and also convert silent resource deletion to diag Warning

* Providers -> ProviderFactories

* ConfigureFunc -> ConfigureContextFunc, use schema.Provider.UserAgent and remove logic for pre-v0.12.0

There was some logic for constructing the User Agent in TF versions earlier than v0.12.0 but those versions are now deprecated and unsupported by the plugin SDK anyway.

Had to use the fmt string for the sweeper User Agent but it should be the same as it was before.

* Add tfproviderlint and fix errors

Doesn't fix all the errors but gets a decent bunch of them.

* Move plural TLS data sources tests to TestCheckTypeSetElemAttrPair

* Add new diag warning when "replace" flag set by Fastly

* WaitForState -> WaitForStateContext

* Check context cancellation in Service Read

* Updates to v2.4.4 of Terraform Plugin SDK

* Updates dependencies with removal of migrate tool and tidy

* Updates TestAccFastlyServiceV1_directors tests to removed unused function parameter

* Removed SDK upgrade comment that is no longer needed

* Updates README.md file to reference 0.26.0 as the last release to support Terraform 0.11

* Updates the backend to be required for the vcl and compute services

Co-authored-by: Ben Gesoff <ben.gesoff@opencredo.com>