Delete StateFuncs from logging blocks
The StateFuncs (which trim the whitespace from the private key and cert attributes), were creating a diff between the config and the state.
This was not an issue before, but when adding the CustomizeDiff function in, and calling d.GetChange(), this difference was returned.
I tried a few things but couldn't get the function to ignore this change without doing some strange heuristic thing trying to match the set elements.
This was mentioned in the comments around the use of the StateFunc already as a known issue with TypeSets, but it didn't seem like there was a way around this while using CustomizeDiff in this way.

Specifically, the unwanted diff in CustomizeDiff was triggering the logic to recompute the service versions (see previous commit) and resulting in a non-empty plan after every apply (and failing the tests as a result).
It should be noted that only the BigQuery tests were failing, even though most of the logging blocks were doing the same thing.
Anyway, I removed the StateFunc logic from all of them and reran the tests and it seems fine.
I've also tested the provider outside of the tests with this change and can't get it to break.