When updating service, clone cloned_version instead of active_version
When making multiple updates to a service with activate=false, we have
been seeing some strange results on the Fastly side of things. This is
because Terraform's view of the current state of a service tracks the
most recent version. When it makes a change, it clones a version and
updates it, so its changes are in the more recent version. However, it
clones the active version each time, so if multiple changes are made,
Terraform's plan is based off the diff between the most recent version
and the desired state, but it applies this plan to the active version
which may be completely different to the most recent version.

This isn't usually a problem when activate=true, because the most recent
version and the active version are always the same. However when the
versions aren't activated each time, there might be a string of draft
versions after the active version, each with an intermediate subset of
the changes. In any case, the version that gets cloned to make the new
version needs to be the version that Terraform used in its plan to
achieve the desired state. The attribute that represents the most
recent, not necessarily activated, version is cloned_version so we
should use this instead.

The acceptance tests aren't yet passing with this change, but the
behaviour is more intuitive in manual testing.