Fix failing tests
The ACL test was previously passing due to the incorrect behaviour from
cloning service versions. The backend name and domain name were being
recalculated in the function for generating the HCL on each test step,
and there are six test steps, so each step also includes a
backend/domain rename as well as the ACL change it is designed to test.
On top of not being the limited-scope test we want, it was also
depending on the buggy behaviour of always cloning the active version.

This came to light in particular when using a step that expects an
error. In this case, the previous step didn't successfully complete, so
the cloned version was never activated. Therefore, when the next step
ran, there was a superfluous domain left in the cloned version which
should have been cleared up. This wasn't a problem previously as the
next version clone would have ignored this version and just cloned the
active version. However fixing this behaviour exposed the test flaw.