Add ACL test step to reset config after ExpectError test step
Without this, the step 5 plan is based on a diff as if step 4  was successful, but it hopefully wasn't, so the diff should essentially be against the config of step 3. This commit adds an extra step to the test case to check the state matches step 3 still, before continuing with the rest of the force_destroy/rename.