Add entry to both ACLs in ACL test
Only one ACL was having an entry added to it. When it came to renaming
the ACLs, half of the time the empty ACL was being renamed first and
succeeding, whereas the other half of the time, the ACL with the entry
was being renamed first and failing thanks to the force_destroy option.
This meant that when checking for ACLs later on, sometimes one of them
had been renamed and couldn't be found. This commit fixes this by adding
an entry to both ACLs so it doesn't matter which is renamed first.