Stop just mutating version 1 when active_version=0
It looks like this logic was added so that when the service is created, the changes are added to version 1 instead of creating the service and immediately cloning to version 2.
This sort of logic is needed due to the fact that the Create function calls the Update function to handle setting all of the nested blocks.
A nice way to handle this would be to extract all of the logic from the Update function into separate, more atomic functions, and then call into them as needed in the Create and Update functions.
This would avoid boolean flag logic all of the Update and Read functions to determine whether we're in Update after Create, normal Update, Read after import, normal Read etc etc.
However this is probably best in another commit/PR.

This commit adds another isCreate flag, in keeping with the temporary convention, to decide whether to skip cloning or not.
In doing so, we can be a bit more explicit about what we want to happen specifically after creating the service, instead of relying on version=0 conditions and having to infer the intent.
Now, when creating, we set cloned_version to 1 in the Create function, and skip cloning straight away with the isCreate flag.

The reason it didn't work to just check active_version=0, is that this condition didn't only occur after creating the service: it could also happen in other scenarios such as importing a service with multiple drafts and no activations.
In this case, even if a service had 20 drafts prior to Terraform managing it, all changes would go as mutations to version 1, which isn't very intuitive at all.