Always active service when activate=true, even if there are no other changes.
Previously, if activate is false then changed to true, there would have to be a change to a nested block as well in order for the service to actually get activated.
This is because the activation logic was in an if block that was only triggered if the "service attribute handler" block function HasChange returned true.
This commit moves the activation logic outside of this if block so that, if activate is true, and the cloned_version is higher than the active_version, then the service will always be activated, regardless of any other changes.