When updating service, clone the cloned_version instead of active_version (#390)
* When updating service, clone cloned_version instead of active_version

When making multiple updates to a service with activate=false, we have
been seeing some strange results on the Fastly side of things. This is
because Terraform's view of the current state of a service tracks the
most recent version. When it makes a change, it clones a version and
updates it, so its changes are in the more recent version. However, it
clones the active version each time, so if multiple changes are made,
Terraform's plan is based off the diff between the most recent version
and the desired state, but it applies this plan to the active version
which may be completely different to the most recent version.

This isn't usually a problem when activate=true, because the most recent
version and the active version are always the same. However when the
versions aren't activated each time, there might be a string of draft
versions after the active version, each with an intermediate subset of
the changes. In any case, the version that gets cloned to make the new
version needs to be the version that Terraform used in its plan to
achieve the desired state. The attribute that represents the most
recent, not necessarily activated, version is cloned_version so we
should use this instead.

The acceptance tests aren't yet passing with this change, but the
behaviour is more intuitive in manual testing.

* Fix failing tests

The ACL test was previously passing due to the incorrect behaviour from
cloning service versions. The backend name and domain name were being
recalculated in the function for generating the HCL on each test step,
and there are six test steps, so each step also includes a
backend/domain rename as well as the ACL change it is designed to test.
On top of not being the limited-scope test we want, it was also
depending on the buggy behaviour of always cloning the active version.

This came to light in particular when using a step that expects an
error. In this case, the previous step didn't successfully complete, so
the cloned version was never activated. Therefore, when the next step
ran, there was a superfluous domain left in the cloned version which
should have been cleared up. This wasn't a problem previously as the
next version clone would have ignored this version and just cloned the
active version. However fixing this behaviour exposed the test flaw.

* Only set cloned_version after cloning, not in every service Read

Avoids the edge case where a version is cloned outside of Terraform, and therefore the most recent version returned by GetServiceDetails might not be the version previously cloned by Terraform and the version that matches Terraform's expectation of the previous state.

* Add ACL test step to reset config after ExpectError test step

Without this, the step 5 plan is based on a diff as if step 4  was successful, but it hopefully wasn't, so the diff should essentially be against the config of step 3. This commit adds an extra step to the test case to check the state matches step 3 still, before continuing with the rest of the force_destroy/rename.

* Add quadratic falloff retry to the ACL attribute check function

In case it gets a 404 and the ACL isn't ready yet, try again up to 10 times, eventually waiting a minute before retrying

* Add entry to both ACLs in ACL test

Only one ACL was having an entry added to it. When it came to renaming
the ACLs, half of the time the empty ACL was being renamed first and
succeeding, whereas the other half of the time, the ACL with the entry
was being renamed first and failing thanks to the force_destroy option.
This meant that when checking for ACLs later on, sometimes one of them
had been renamed and couldn't be found. This commit fixes this by adding
an entry to both ACLs so it doesn't matter which is renamed first.

* Rework service resource read logic for choosing version

Now sets cloned_version during import. Also always reads cloned_version
when activate is false. Previously it would always read the most recent
version but now it will read the version last cloned by the provider.

Also a few random comment/description changes.

* Reuse initial version if no version has been activated

If no version has been activated, do not clone and edit that, just use version 1 for edits. Mainly affects the import flow where cloned_version gets imported to the latest version.

* Revert af2a6c8 and 45bf459 (it was something else)

These commits were supposed to fix the ACL tests but didn't. In the end
the fix was 5b45f0c so these commits are unnecessary.

* Stop just mutating version 1 when active_version=0

It looks like this logic was added so that when the service is created, the changes are added to version 1 instead of creating the service and immediately cloning to version 2.
This sort of logic is needed due to the fact that the Create function calls the Update function to handle setting all of the nested blocks.
A nice way to handle this would be to extract all of the logic from the Update function into separate, more atomic functions, and then call into them as needed in the Create and Update functions.
This would avoid boolean flag logic all of the Update and Read functions to determine whether we're in Update after Create, normal Update, Read after import, normal Read etc etc.
However this is probably best in another commit/PR.

This commit adds another isCreate flag, in keeping with the temporary convention, to decide whether to skip cloning or not.
In doing so, we can be a bit more explicit about what we want to happen specifically after creating the service, instead of relying on version=0 conditions and having to infer the intent.
Now, when creating, we set cloned_version to 1 in the Create function, and skip cloning straight away with the isCreate flag.

The reason it didn't work to just check active_version=0, is that this condition didn't only occur after creating the service: it could also happen in other scenarios such as importing a service with multiple drafts and no activations.
In this case, even if a service had 20 drafts prior to Terraform managing it, all changes would go as mutations to version 1, which isn't very intuitive at all.

* Always active service when activate=true, even if there are no other changes.

Previously, if activate is false then changed to true, there would have to be a change to a nested block as well in order for the service to actually get activated.
This is because the activation logic was in an if block that was only triggered if the "service attribute handler" block function HasChange returned true.
This commit moves the activation logic outside of this if block so that, if activate is true, and the cloned_version is higher than the active_version, then the service will always be activated, regardless of any other changes.