Catch case where state from older version could be unexpected (#396)
A change in v0.28.1 moved to use cloned_version much more heavily. This
made a few assumptions that during import/create, the cloned_version
would be set, and that therefore in a usual Read, this would already be
taken care of. However this missed a situation where the user has an
existing state file from v0.28.0 or earlier and upgrades to v0.28.1. In
this situation, neither the Create nor the Import logic would be
triggered and the cloned_version would have evaded attention. This then
causes problems in the Update function where we assume cloned_version is
correct and use it to clone a new version to make changes. In particular
this commit addresses the case where cloned_version is 0, which leads to
an error message when calling CloneVersion:

Error: missing required field 'ServiceVersion'

To get around this, we look for the case where cloned_version is 0 in
the Read function and use the same logic as Import in that case too.