Add DiffSuppressFunc to optionally suppress all diffs on ACL entries
Adds a manage_entries attribute which defaults to false. Then a
DiffSuppressFunc reads this and decides whether to suppress the diff on
the entries set. By default this then only applies the changes on create
and does nothing the rest of the time, unless the user sets
manage_entries to true manually, then it behaves as usual. This is to
make the resource safer by default in the event that entries are added
to the ACL outside of Terraform, so that they aren't accidentally
deleted.