Update default initial value for health check.
Takes into account the change where the VCL is lazy-loaded on nodes as
needed.  The lazy-load may cause the first couple of health check
attempts to fail, and can be enough to mark the backend as unhealthy
until enough successful checks occur to reach the threshold again.

It is highly recommended that the initial and threshold values be the
same for health checks going forward.

By setting the initial value to equal the threshold value, the threshold
will be met immediately, before any Health check requests are sent.  The
backend is initialized as Healthy from the start. This avoids
unnecessary errors as the VCL is being loaded onto the nodes.