Fix: only set cloned_version after the version has been successfully validated
Also adds a test case which previously failed with a 409 error prior to this fix.

Beforehand, if there was an error midway through an update, and `activate` was true, there would be a half-updated draft version left on the service, which wouldn't be activated.
The next apply would then read the active version but clone this half-updated version and try to apply the plan to it.
This could cause a 409 error in some cases as the provider would be applying changes to a different version from which it had read.

The fix for this is to only set `cloned_version` after the version validation has successfully passed.
This means that the next apply after an unsuccessful one will clone from the active version, as expected.