When `activate=true`, always read and clone from the active version
Before, updates would all be based off `cloned_version`, and `cloned_version` would always be the last version that Terraform made changes to.
However, the version that the state was read from depended on whether `activate` was `true`.
If so, then state was read from the `active version`, and it was assumed that this would be the same as `cloned_version` as `activate` was `true` and the `cloned_version` was always activated after it was created.
If `activate` was `false`, then the state would be read from `cloned_version`.

However, this assumption was incorrect.
It is valid when `activate` is `false`, but when `activate` is `true`, it does not account for the fact that the version could be created and activated outside of Terraform.
In this case, the version that state is read from, and the one the updates are applied to, can be different, and errors can occur.

This commit changes the behaviour so that, when `activate` is `true`, the value of `cloned_version` is updated on every state read to match the currently active version.
This ensures that the above assumption is valid, even when a version is cloned and activated outside of Terraform.

There is also an acceptance test added to replicate the behaviour that led to original bug, to validate that it has been fixed.

With this fix in place, it should now always be true that the same version will be used for reading the state and making updates to.