Improve error message when importing non-existent service by returning the 404 instead of nil
Previously the error said:

Error: nil entry in ImportState results. This is always a bug with
the resource that is being imported. Please report this as
a bug to Terraform.

This was because a 404 on a service read would call `d.SetId("")` and then return `nil`.
To fix this, I just returned the error instead of `nil`, which meant the terminal output is now:

Error: 404 - Not Found
Cannot find service 'asfasd'