In fastly_waf_rules data source, request rule revisions from API
To populate the Revisions field of the rules returned from ListAllWAFRules, the Include field of the request must specify "waf_rule_revisions".
Otherwise this field is always nil.
This commit just adds "waf_rule_revisions" to the request.

Furthermore there were some tests for the same data source which don't match the naming convention for acceptance tests and unit tests.