Simplify service importer and allow specifying version to import
Previously the service resource importer was doing a full state refresh.
This was redundant because the Read function gets called right after the Import one, so the state was being refreshed twice.
It also introduced extra logic in the Read function to determine whether it was being called via the Import function or not, to enable the version selection logic to be tweaked.

This commit changes the behaviour so that the Import function only has one responsibility; to set cloned_version.
The rest of the state read is then fully delegated to the Read function, which needs no extra logic to determine the version, as it should already be set.

The Import function has been given some extra logic, to optionally parse a version number from the provided ID, by using an '@' symbol after the ID.
If it isn't specified, i.e. only the ID is used, then the Read function will default to the active version, or the latest version if that is not set either.