Add tool which auto-generates AST for refactoring common `SetDiff` files
The last commit manually refactored about three of the blocks, but there are way too many blocks to think about doing all of them in this way.
Therefore I have started a tool to do this automatically.

It uses the `go/ast` package to parse the source files and manipulate them into the new format.
Currently it just outputs the results to stdout to ease development, rather than modifying them in place.