V1 versionless deletion protection (#110)
* Add DiffSuppressFunc to optionally suppress all diffs on ACL entries

Adds a manage_entries attribute which defaults to false. Then a
DiffSuppressFunc reads this and decides whether to suppress the diff on
the entries set. By default this then only applies the changes on create
and does nothing the rest of the time, unless the user sets
manage_entries to true manually, then it behaves as usual. This is to
make the resource safer by default in the event that entries are added
to the ACL outside of Terraform, so that they aren't accidentally
deleted.

* Add acceptance tests for manage_entries attribute

* Add DiffSuppressFunc manage_entries to dictionary items as well

Same change as for the ACL entries resource, but for the dictionary
items.

* Run make generate-docs

* add a missing argument

* re-generate docs

Co-authored-by: Ben Gesoff <ben.gesoff@opencredo.com>
Co-authored-by: Hieu Doan <hiieu@users.noreply.github.com>