Changes for v1.0.0 (#532)
* V1 versionless deletion protection (#110)

* Add DiffSuppressFunc to optionally suppress all diffs on ACL entries

Adds a manage_entries attribute which defaults to false. Then a
DiffSuppressFunc reads this and decides whether to suppress the diff on
the entries set. By default this then only applies the changes on create
and does nothing the rest of the time, unless the user sets
manage_entries to true manually, then it behaves as usual. This is to
make the resource safer by default in the event that entries are added
to the ACL outside of Terraform, so that they aren't accidentally
deleted.

* Add acceptance tests for manage_entries attribute

* Add DiffSuppressFunc manage_entries to dictionary items as well

Same change as for the ACL entries resource, but for the dictionary
items.

* Run make generate-docs

* add a missing argument

* re-generate docs

Co-authored-by: Ben Gesoff <ben.gesoff@opencredo.com>
Co-authored-by: Hieu Doan <hiieu@users.noreply.github.com>

* Make backend `auto_loadbalance` default to false (#111)

The Fastly web interface defaults "Auto load balance" to "No". The most
common reason for having multiple backends in a single service is to
route different paths to different backends, rather than load balance
between different origins. Currently, this provider defaults
`auto_loadbalance` to true, which is surprising and confusing. This
commit makes `auto_loadbalance` default to false to reduce confusion.

Co-authored-by: Ben Gesoff <ben.gesoff@opencredo.com>
Co-authored-by: Hieu Doan <hiieu@users.noreply.github.com>

* V1 update default format and format version (#112)

* Update default values for format & format_version

* Update format & format_version in docs and doc strings

* Cosmetic change: capitilise Default

* Update format & format_version in tests

* Fix 'format' for papertail

* add FormatVersion for the GCS struct

* update gcslogging doc with format_version

* change double quotes to backticks

Co-authored-by: Hieu Doan <hiieu@users.noreply.github.com>

* V1 change email to user in gcslogging (#113)

* Change the email field to user to align with the API docs.

* fix TestGcsloggingEnvDefaultFuncAttributes by using the user attribute

* fix TestResourceFastlyFlattenGCS

* go fmt

Co-authored-by: Hieu Doan <hiieu@users.noreply.github.com>

* use `TypeList` schema for `gzip` nested block attributes (#114)

Co-authored-by: Shohei Maeda <irt_m.jrsyo@ntworkers.com>
Co-authored-by: Hieu Doan <hiieu@users.noreply.github.com>

* Consistent resource naming (#115)

Consistent resource naming:
- change `v1` to `vcl`
- remove `v1` 
- add `logging` prefix
- update docs and examples

Other changes:
- fix acl manage entries test
- fix `user` attribute in gcs

Co-authored-by: Hieu Doan <hiieu@users.noreply.github.com>

Co-authored-by: Ben Gesoff <ben.gesoff@opencredo.com>
Co-authored-by: Hieu Doan <hiieu@users.noreply.github.com>
Co-authored-by: Shohei Maeda <irt_m.jrsyo@ntworkers.com>