Support Service Authorizations (#572)
* Add support for Service Authorizations

This change adds support for [Fastly Service Authorizations][1] to be
managed via Terraform.

Other managed users and services can be referenced by their ID and the
permission is controlled with an enumeration of the allowed values by
the API (`full`, `read_only`, `purge_select` or `purge_all`).

[1]: https://developer.fastly.com/reference/api/account/service-authorization/

* Service Authorizations: Adds docs

Add documentation on how to use and import services authorizations with
Fastly.

* service_authorization: Align demo resource names

* service_authorization: Require permission

The default permission in the API is `full` which is extremely
permissive and arguably not a good default. The terraform provider will
require the permission to be set to avoid over-permitting users by a
simple omission.

* go.mod: Bump version of go-fastly

* service_authorization: Add acceptance test

Acceptance tests ensure the resource works as expected right now, but
also that the functionality keeps working as other changes are applied.

* service_authorization: Re-generate docs

* Use main branch of fastly/go-fastly

Removes the temporary replace directive. This should still be replaced
with a tag once the changes on main have actually been released.

* go.mod: Use released version of github.com/fastly/go-fastly/v6