data source: Fastly Services
This adds a new datasource for listing service ids, and details about
services.

This can be used if you need to read the service configurations to
configure some other resource. For instance it could be useful to
combine this with a `fastly_service_authorization` to grant permissions
to all services in an account.