remove new test to validate env var default...
I removed this test because although it worked locally, once running the full
test suite I quickly discovered that the pausing of a test (to be resumed later,
i.e. concurrent processing) was causing the environment to get very confused and
it wasn't adding a whole lot of value to be honest.