Fix tfproviderlint errors (#596)
* replace interface{} with any

* handle errors

* remove dead code

* remove underscores from test names

* Revert "remove underscores from test names"

This reverts commit 6d6473dd5beb422b4cea6db215af642dbdc95bdf.

* remove import from test name

* skip XAT001 missing resource.TestCase ErrorCheck

ErrorCheck allows providers the option to handle errors such as skipping tests
based on certain errors. But we don't need to worry about that for our
current test suite. Maybe in the future we'll want to add an ErrorCheck
but we shouldn't feel like it's mandatory.

https://github.com/bflad/tfproviderlint/tree/main/xpasses/XAT001
https://pkg.go.dev/github.com/hashicorp/terraform-plugin-sdk/v2@v2.23.0/helper/resource#TestCase

* fix AT003 to ensure 'per resource testing' is made easier

* fix XS002 (ordered keys)

* skip R001 for complex d.Set() calls

We don't want to have to manually type the key name when we already have
h.GetKey()/h.Key() to return us the value. Especially if we decide to
rename that key in the future.

https://github.com/bflad/tfproviderlint/tree/main/passes/R001

* fix AT002

* fix AT003

* fix AT004

* fix AT012

* skip R018

* tfproviderlintx

* ignore lint errors via CLI

* Revert "skip R001 for complex d.Set() calls"

This reverts commit add29c44cf3857265ee839ee6953c2e9807f02fd.

* Revert "skip XAT001 missing resource.TestCase ErrorCheck"

This reverts commit 0d1874640f02f5c1afd214a76d30269bbd8fde06.

* fix makefile